refactor(services): replace generateJobName function with anthropicClient integration for improved modularity
